add clickhouse and kafka sink support.

+++ src/main/java/com/github/capture/model/TcpPacketRecord.java
@@ -36,7 +36,7 @@ public class TcpPacketRecord implements Serializable {
 
     private byte ipTos;
 
-    private short ipTotalLen;
+    private int ipTotalLen;
 
     private int ipIdentification;
